breaking sql and backend function into their own files and adding head file
